requirements.txt
aiogram~=3.18.0
python-dotenv~=1.0.1
beautifulsoup4~=4.13.3
requests~=2.32.3
